Update overview to focus more on UX of sections
If this is the first we introduce someone to the different areas of the 
Block Editor I alrready want to bring across what kind of settings
should be placed in the different sections of the editor.

This point should be expanded uplon further in following sections where
it gets explained which settings should go where and why.

The Block Design guide in the Gutenberg Docs is a great ressource for
this: https://github.com/WordPress/gutenberg/blob/trunk/docs/how-to-guides/designers/block-design.md